Multiple breaking changes
*) Refactoring of REL_ONE/REL_MANY to RelOne/RelMany
*) DocumentBase renamed to DocumentModel
The DocumentModel struct will be used to apply the model
concept to a generic struct. Using the model concept we
can add meta information directly inside the struct using
field tags just like "bson" and "json" tags.

-- Under Development --

*) Introduction of the coll tag and the GetCollectionName to
extract the collection name that should be used to make the
mongo collection

*) Introduction of the index tag and an index parser (the golang base
parser in the standard library) to be used with the mgo/Index struct
and mgo/EnsureIndex method to make indexes related to the document model
struct.

...