Introducing the Model interface
The Model interface main purpose is  to enrich the concept of model of
the document. This interface gives access to the meta information which
are not strictly related to the data contained in the document.
This interface is used to embed within to the model of the document meta
information used by the mgo driver to, for instance, select the target
collection or the indexes to be built for the document.

Also the NewDocumentModel() function is added to permit a cleaner
way to make a new document model, and to have a simplified access
to its fields.