Added the me field
The me field of DocumentModel is a pointer to the document itself and
it will be used to have a reference to the document type when needed.

Using this field we can bypass the Go access restriction to the document
data when calling a receiver of DocumentModel. In this way we can
make a call like doc.Save(), without the need to pass the doc itself
as generic interface{}.

Also the following receivers are added:
AsModel
AsDocument
GetMe - used in find
SetMe - used in find to set the me field
Find
FindID
Save
Remove