Changed namespace / Population process
The namespace changed from Bongo to Mogo, togheter with the name of the
project. This because i made an heavy refactoring to the whole project
which makes it very different from the previous one.

This commit also contains methods to implement the population process,
which is the process of automatically replacing the referenced field
pointing to other documents in other collections.